pyproject.toml
[tool.poetry]
name = "hotel-vacancy"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "3.9.15"
pynamodb = "^5.3.4"
pydantic = "^1.10.4"
PyYAML = "^6.0"
slackweb = "^1.0.5"
Jinja2 = "^3.1.2"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
isort = "^5.11.4"
mypy = "^0.991"
black = "^22.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
